file reordering, to extend graphql types for future use cases
