--- src/client.ts
@@ -1,18 +1,0 @@
-import { ApolloClient, HttpLink, InMemoryCache } from '@apollo/client/core';
-import fetch from 'cross-fetch';
-import config from './config';
-
-export const client = new ApolloClient({
-  link: new HttpLink({ fetch, uri: config.clientApi.uri }),
-  cache: new InMemoryCache(),
-  name: config.app.apolloClientName,
-  version: config.app.version,
-  defaultOptions: {
-    watchQuery: {
-      fetchPolicy: 'no-cache',
-    },
-    query: {
-      fetchPolicy: 'no-cache',
-    },
-  },
-});
